fix (GoBackIcon): fixed button at bottom left
